+++ src/openrct2-ui/drawing/engines/opengl/TextureCache.cpp
@@ -18,7 +18,7 @@
     #include <openrct2/drawing/Drawing.h>
     #include <openrct2/drawing/FilterPaletteIds.h>
     #include <openrct2/drawing/RenderTarget.h>
-    #include <openrct2/interface/Screen.hpp>
+    #include <openrct2/interface/ScreenCoords.hpp>
     #include <stdexcept>
     #include <vector>
