+++ src/openrct2-ui/input/InputManager.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include <openrct2/interface/Screen.hpp>
+#include <openrct2/interface/ScreenCoords.hpp>
 #include <queue>
 #include <string_view>
