+++ src/openrct2-ui/interface/Graph.h
@@ -12,7 +12,7 @@
 #include <openrct2/Context.h>
 #include <openrct2/core/Money.hpp>
 #include <openrct2/interface/ColourWithFlags.h>
-#include <openrct2/interface/Screen.hpp>
+#include <openrct2/interface/ScreenCoords.hpp>
 
 namespace OpenRCT2::Drawing
 {
