+++ src/openrct2/interface/ScreenCoords.hpp
