+++ src/openrct2/object/EntranceObject.cpp
@@ -13,7 +13,7 @@
 #include "../core/IStream.hpp"
 #include "../core/Json.hpp"
 #include "../drawing/Drawing.h"
-#include "../interface/Screen.hpp"
+#include "../interface/ScreenCoords.hpp"
 #include "../localisation/Language.h"
 #include "../world/tile_element/EntranceElement.h"
