+++ src/openrct2/object/PathAdditionObject.cpp
@@ -14,7 +14,7 @@
 #include "../core/Json.hpp"
 #include "../drawing/Drawing.h"
 #include "../interface/Cursors.h"
-#include "../interface/Screen.hpp"
+#include "../interface/ScreenCoords.hpp"
 
 namespace OpenRCT2
 {
