+++ src/openrct2/object/SmallSceneryObject.cpp
@@ -16,7 +16,7 @@
 #include "../core/Json.hpp"
 #include "../drawing/Drawing.h"
 #include "../interface/Cursors.h"
-#include "../interface/Screen.hpp"
+#include "../interface/ScreenCoords.hpp"
 #include "../localisation/Language.h"
 
 namespace OpenRCT2
