+++ src/openrct2/object/TerrainEdgeObject.cpp
@@ -13,7 +13,7 @@
 #include "../core/Guard.hpp"
 #include "../core/Json.hpp"
 #include "../drawing/Drawing.h"
-#include "../interface/Screen.hpp"
+#include "../interface/ScreenCoords.hpp"
 #include "ObjectManager.h"
 
 namespace OpenRCT2
