+++ src/openrct2/object/TerrainSurfaceObject.cpp
@@ -15,7 +15,7 @@
 #include "../core/Guard.hpp"
 #include "../core/Json.hpp"
 #include "../drawing/Drawing.h"
-#include "../interface/Screen.hpp"
+#include "../interface/ScreenCoords.hpp"
 #include "../world/Location.hpp"
 #include "ObjectManager.h"
