+++ src/openrct2/object/WallObject.cpp
@@ -15,7 +15,7 @@
 #include "../drawing/Drawing.h"
 #include "../drawing/ScrollingText.h"
 #include "../interface/Cursors.h"
-#include "../interface/Screen.hpp"
+#include "../interface/ScreenCoords.hpp"
 #include "../localisation/Language.h"
 
 namespace OpenRCT2
